requirements.txt
# Core dependencies
llama-index>=0.12.0
python-dotenv==1.0.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
chromadb>=0.5.17
pypdf>=5.1.0
langchain==0.1.9
sqlalchemy==2.0.27
pydantic>=2.0.0
fastmcp==0.4.1